Add createAndSendUpdateCommitment to reducer-helpers
